refactor: rename getTagName to getKaioTagName and abstract traversing
