+++ packages/components/codemods/utils/getKaioTagName.spec.ts
@@ -1,10 +1,10 @@
 import { parseJsx } from "../__tests__/utils"
-import { getTagName } from "./getTagName"
+import { getKaioTagName } from "./getKaioTagName"
 
-describe("getTagName", () => {
+describe("getKaioTagName", () => {
   it("returns the import name if it matches the target specifier", () => {
     const input = parseJsx('import { Well } from "@kaizen/components"')
-    const tagName = getTagName(input, "Well")
+    const tagName = getKaioTagName(input, "Well")
     expect(tagName).toBe("Well")
   })
 
@@ -12,13 +12,13 @@ describe("getTagName", () => {
     const input = parseJsx(
       'import { Well as KaizenWell } from "@kaizen/components"'
     )
-    const tagName = getTagName(input, "Well")
+    const tagName = getKaioTagName(input, "Well")
     expect(tagName).toBe("KaizenWell")
   })
 
   it("returns undefined if there is no match to the target specifier", () => {
     const input = parseJsx('import { Well } from "@kaizen/well"')
-    const tagName = getTagName(input, "Well")
+    const tagName = getKaioTagName(input, "Well")
     expect(tagName).toBe(undefined)
   })
 })
